Unit tests ok, integration test works
